style: :art: add norm and pathing organization
