+++ minishell/Makefile
@@ -5,16 +5,18 @@ CFLAGS = -Wall -Wextra -Werror
 
 SRC_DIR = src
 LIB_DIR = lib
+BUILTIN_DIR = builtins
 
 LIBFT = $(LIB_DIR)/libft.a
 
 SRCS = \
 	$(SRC_DIR)/main.c \
-	$(SRC_DIR)/utilits.c \
-	$(SRC_DIR)/builtins.c \
-	$(SRC_DIR)/builtins2.c \
-	$(SRC_DIR)/frees.c \
-	$(SRC_DIR)/error.c \
+	$(SRC_DIR)/$(BUILTIN_DIR)/utilits.c \
+	$(SRC_DIR)/$(BUILTIN_DIR)/utilits_env.c \
+	$(SRC_DIR)/$(BUILTIN_DIR)/utilits_export.c \
+	$(SRC_DIR)/$(BUILTIN_DIR)/builtins.c \
+	$(SRC_DIR)/$(BUILTIN_DIR)/builtins2.c \
+	$(SRC_DIR)/$(BUILTIN_DIR)/error.c \
 
 OBJS = $(SRCS:.c=.o)
